+++ functions.php
@@ -2,7 +2,7 @@
 /**
  * Functions and definitions
  *
- * @package Penguin Gold
+ * @package Penguin
  */
 
 if ( ! function_exists( 'penguin_setup' ) ) :
@@ -19,7 +19,6 @@ function penguin_setup() {
 	 * Make theme available for translation.
 	 */
 	load_theme_textdomain( 'penguin' );
-	load_theme_textdomain( 'penguin-gold', get_template_directory() . '/languages' );
 
 	// Add default posts and comments RSS feed links to head.
 	add_theme_support( 'automatic-feed-links' );
@@ -122,13 +121,12 @@ add_action( 'widgets_init', 'penguin_widgets_init' );
  */
 function penguin_scripts() {
 	$theme     = wp_get_theme();
-	$penguin   = wp_get_theme( 'penguin-gold' );
 	$minified  = get_theme_mod( 'min-files' );
 	$suffix    = ( 1 === $minified ) ? '.min' : '';
 	$fluidvids = get_theme_mod( 'fluidvids', true );
 
 	if ( is_child_theme() ) {
-		wp_enqueue_style( 'penguin-parent-style', get_template_directory_uri() . '/style.css', false, $penguin['Version'] );
+		wp_enqueue_style( 'penguin-parent-style', get_template_directory_uri() . '/style.css', false, $theme['Version'] );
 		wp_style_add_data( 'penguin-parent-style', 'suffix', $suffix );
 	}
 
@@ -139,8 +137,9 @@ function penguin_scripts() {
 		wp_enqueue_script( 'penguin-masonry-options', get_template_directory_uri() . '/js/masonry-options' . $suffix . '.js', array( 'masonry' ), '1.0', true );
 	}
 
-	wp_enqueue_script( 'smooth-scroll', get_template_directory_uri() . '/js/smooth-scroll' . $suffix . '.js', array(), '5.3.3', true );
+	wp_enqueue_script( 'smooth-scroll', get_template_directory_uri() . '/js/smooth-scroll' . $suffix . '.js', array(), '10.2.1', true );
 	wp_enqueue_script( 'penguin-navigation', get_template_directory_uri() . '/js/navigation' . $suffix . '.js', array(), '20161106', true );
+	wp_enqueue_script( 'svg4everybody', get_template_directory_uri() . '/js/svg4everybody' .$suffix . '.js', array(), '2.1.4', true );
 
 	if ( 1 == $fluidvids ) {
 		wp_enqueue_script( 'fluidvids', get_template_directory_uri() . '/js/fluidvids' . $suffix . '.js', array(), '2.4.1', true );
@@ -165,20 +164,20 @@ function penguin_show_custom_image_sizes( $sizes ) {
 add_filter( 'image_size_names_choose', 'penguin_show_custom_image_sizes' );
 
 /**
- * Add custom image sizes attribute to enhance responsive image functionality 
+ * Add custom image sizes attribute to enhance responsive image functionality
  * for content images
  *
  * @since Penguin 0.2
  * @return string A source size value for use in a content image 'sizes' attribute.
  */
 function penguin_content_image_sizes_attr($size) {
-	// Singular posts with sidebar
-	if ( is_singular() || is_sticky() ) {
-		return '(max-width: 599px) calc(100vw - 50px), (max-width: 767px) calc(100vw - 70px), (max-width: 991px) 429px, (max-width: 1199px) 637px, 747px';
-	}
 	// Page full width without sidebar
-	if ( get_page_template_slug() === 'page-fullwidth.php' ) {
+	if ( is_page_template('page-templates/page-fullwidth.php' ) || is_page_template('page-templates/page-transparent-fullwidth.php' ) ) {
 		return '(max-width: 599px) calc(100vw - 50px), (max-width: 767px) calc(100vw - 70px), (max-width: 991px) 679px, (max-width: 1199px) 879px, 1039px';
+	}
+	// Singular posts with sidebar
+	elseif ( is_singular() || is_sticky() ) {
+		return '(max-width: 599px) calc(100vw - 50px), (max-width: 767px) calc(100vw - 70px), (max-width: 991px) 429px, (max-width: 1199px) 637px, 747px';
 	}
 	// 2 col blog with sidebar
 	else {
@@ -229,7 +228,9 @@ function penguin_wp_footer() {
 			players: ['www.youtube.com', 'www.youtube-nocookie.com', 'player.vimeo.com']
 		});
 	</script>
-	<?php }
+	<?php } ?>
+	<script>svg4everybody();</script>
+	<?php
 }
 add_action( 'wp_footer', 'penguin_wp_footer', 21 );
 
@@ -242,7 +243,6 @@ require get_template_directory() . '/inc/template-tags.php';
  * Penguin extras
  */
 require get_template_directory() . '/inc/extras-penguin.php';
-require get_template_directory() . '/inc/extras-penguin-gold.php';
 
 /**
  * Plugins compatibility file
@@ -254,8 +254,7 @@ require get_template_directory() . '/inc/extras-plugins.php';
  */
 require get_template_directory() . '/inc/style-builder.php';
 require get_template_directory() . '/inc/customizer-options.php';
-require get_template_directory() . '/inc/customizer-options-gold.php';
-require get_template_directory() . '/inc/customizer-styles-gold.php';
+require get_template_directory() . '/inc/customizer-styles.php';
 
 /*
  * Theme Hook Alliance files